Refactor station info into its own class
To support icon loading I was adding more parameters to
a couple of routines and decided to refactor everything into
a new class that manages the state for a station.

To support the new object added a couple of lookup routines.

Currently there is a bug where the first menu item in the main
menu doesn't show up until I've clicked on the media menu
(where it shows up the first time that I click).