Include all stations and support El Capitain
Due to OS-X 10.11 App Strict transport this app was no longer
building.  This now excludes the icecast servers from that requirement.

Additionally, this adds *all* of the SomaFM stations as of 27-Mar-2016.
The code to add the stations and the images for the stations are
generated by `Utils/fetcher.py`